requirements.txt
pandas
faiss-cpu
sentence-transformers
together
fastapi 
uvicorn 
pydantic
streamlit
bs4
dotenv